requirements.txt
pydantic_ai
pymupdf
requests
beautifulsoup4

